# cmd/app/docs/swagger.yaml
basePath: /
definitions:
...
  financial_system_internal_domain.Enterprise:
    properties:
      balance:
        type: number
      id:
        type: integer
      name:
        type: string
    type: object
  financial_system_internal_domain.EnterpriseWithEmployees:
    properties:
      employee_user_ids:
        items:
          type: integer
        type: array
      enterprise:
        $ref: '#/definitions/financial_system_internal_domain.Enterprise'
    type: object
...
  financial_system_internal_domain.Transaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      from_account_id:
        type: integer
      from_deposit_id:
        type: integer
      id:
        type: integer
      to_account_id:
        type: integer
      to_deposit_id:
        type: integer
      transaction_type:
        type: string
    type: object
...
  internal_transport_rest.addEmployeeToEnterpriseRequest:
    properties:
      user_id:
        example: 3
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления финансами. Клиент, менеджер и администратор.
  title: Financial System API
  version: "1.0"
paths:
...
  /manager/accounts/{id}/block:
    post:
      parameters:
      - description: ID счёта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Заблокировать счёт
      tags:
      - manager
  /manager/accounts/{id}/unblock:
    post:
      parameters:
      - description: ID счёта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Разблокировать счёт
      tags:
      - manager
  /manager/deposits/{id}/block:
    post:
      description: Менеджер может заблокировать вклад в любой момент (без проверки
        на баланс).
      parameters:
      - description: ID вклада
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Заблокировать вклад
      tags:
      - manager
  /manager/deposits/{id}/unblock:
    post:
      parameters:
      - description: ID вклада
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Разблокировать вклад
      tags:
      - manager
  /manager/enterprises:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/financial_system_internal_domain.EnterpriseWithEmployees'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Предприятия с сотрудниками
      tags:
      - manager
  /manager/enterprises/{enterprise_id}/employees/{user_id}:
    delete:
      description: Pending заявки этого сотрудника по предприятию автоматически отклоняются.
      parameters:
      - description: ID предприятия
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить сотрудника из предприятия
      tags:
      - manager
  /manager/enterprises/{id}/employees:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID предприятия
        in: path
        name: id
        required: true
        type: integer
      - description: user_id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_rest.addEmployeeToEnterpriseRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить сотрудника в предприятие
      tags:
      - manager
  /manager/salary-project/applications/{id}/approve:
    post:
      description: Одобряет заявку (status = approved). Баланс предприятия должен
        быть не меньше суммы заявки.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Подтвердить заявку на зарплатный проект
      tags:
      - manager
  /manager/users/{id}/approve:
    post:
      description: Устанавливает is_active = true для пользователя с ролью client.
        После этого клиент может войти по SignIn.
      parameters:
      - description: ID пользователя (клиента)
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Подтвердить регистрацию клиента
      tags:
      - manager
  /manager/users/{id}/history:
    get:
      description: Объединённая история по всем счетам пользователя (аналогично /client/accounts/history
        по каждому счёту), отсортировано по дате.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/financial_system_internal_domain.Transaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: История операций пользователя
      tags:
      - manager
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"